move Model to model.go
split/rename the rest of device_asserts.go to serial_asserts.go
